requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
imapclient
pyzmail36
loguru
